% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_states.R
\name{assign_state}
\alias{assign_state}
\title{Assign states to genes}
\usage{
assign_state(seg, column, patient, ref)
}
\arguments{
\item{seg}{data.frame of cnv}

\item{column}{column name for patients}

\item{patient}{string of patient to keep}

\item{ref}{data.frame of genes}
}
\value{
A data.frame of assigned cnvs
}
\description{
Assign states to genes
}
\examples{
\dontrun{
library(cassign)

res <- assign_state("A", "pat", y, x)
}

}

