% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_states.R
\name{assign_state}
\alias{assign_state}
\title{Assign states to genes}
\usage{
assign_state(patient, P, ref, seg)
}
\arguments{
\item{patient}{string of patient to keep}

\item{P}{column name for patients}

\item{ref}{data.frame of genes}

\item{seg}{data.frame of cnv}
}
\value{
A data.frame of assigned cnvs
}
\description{
Assign states to genes
}
\examples{
\dontrun{
library(cassign)

# Import file
e <- read.plate("od_measure.xls")
e <- elisa.analyze(e)
e <- elisa.analyze(e, blank = TRUE, transform = TRUE)
}

}

